refactor: `std::size_t` -> `std::uint64_t` for ids

+++ include/saucer/serializers/data.hpp
@@ -6,7 +6,7 @@ namespace saucer
 {
     struct message_data
     {
-        std::size_t id;
+        std::uint64_t id;
         virtual ~message_data() = default;
     };
